# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LlmConsentOut": {
        "properties": {
          "confirmed_at": {
            "format": "date-time",
            "title": "Confirmed At",
            "type": "string"
          },
          "confirmed_by_name": {
            "title": "Confirmed By Name",
            "type": "string"
          }
        },
        "required": [
          "confirmed_by_name",
          "confirmed_at"
        ],
        "title": "LlmConsentOut",
        "type": "object"
      },
      "LlmConsentRequest": {
        "properties": {
          "confirmed_by_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Confirmed By Name",
            "type": "string"
          }
        },
        "required": [
          "confirmed_by_name"
        ],
        "title": "LlmConsentRequest",
        "type": "object"
      },
...
      "MemberIn": {
        "properties": {
          "role": {
            "enum": [
              "owner",
              "editor",
              "viewer",
              "client"
            ],
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "role"
        ],
        "title": "MemberIn",
        "type": "object"
      },
      "MemberOut": {
        "properties": {
          "account_type": {
            "title": "Account Type",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "display_name",
          "account_type",
          "role"
        ],
        "title": "MemberOut",
        "type": "object"
      },
      "ProjectCreate": {
        "properties": {
          "client_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "name": {
            "maxLength": 200,
            "title": "Name",
            "type": "string"
          },
          "storage_connection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Connection Id"
          },
          "storage_root": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Root"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectOut": {
        "properties": {
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "llm_consent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LlmConsentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "my_role": {
            "title": "My Role",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "storage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStorageOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "client_name",
          "created_at",
          "my_role",
          "settings",
          "storage",
          "llm_consent"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
      "ProjectSettings": {
        "additionalProperties": false,
        "properties": {
          "check_budget_usd": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "maximum": 50.0,
            "title": "Check Budget Usd",
            "type": "number"
          },
          "model": {
            "default": "claude-opus-5",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "normalize_budget_usd": {
            "default": 2.0,
            "exclusiveMinimum": 0.0,
            "maximum": 50.0,
            "title": "Normalize Budget Usd",
            "type": "number"
          }
        },
        "title": "ProjectSettings",
        "type": "object"
      },
      "ProjectSettingsPatch": {
        "additionalProperties": false,
        "properties": {
          "check_budget_usd": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 50.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Budget Usd"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "normalize_budget_usd": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 50.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalize Budget Usd"
          }
        },
        "title": "ProjectSettingsPatch",
        "type": "object"
      },
      "ProjectStorageOut": {
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "connection_name": {
            "title": "Connection Name",
            "type": "string"
          },
          "root": {
            "title": "Root",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "connection_name",
          "type",
          "root"
        ],
        "title": "ProjectStorageOut",
        "type": "object"
      },
      "ProjectUpdate": {
        "properties": {
          "client_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectSettingsPatch"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ProjectUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QC-Agent",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "title": "Response List Projects Api V1 Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "operationId": "archive_project_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/llm-consent": {
      "post": {
        "operationId": "record_llm_consent_api_v1_projects__project_id__llm_consent_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmConsentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Llm Consent",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/members": {
      "get": {
        "operationId": "get_members_api_v1_projects__project_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberOut"
                  },
                  "title": "Response Get Members Api V1 Projects  Project Id  Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Members",
        "tags": [
          "projects"
        ]
      },
      "put": {
        "operationId": "put_members_api_v1_projects__project_id__members_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/MemberIn"
                },
                "title": "Members",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberOut"
                  },
                  "title": "Response Put Members Api V1 Projects  Project Id  Members Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Members",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}